+++ src/modules/geo/GoActive.Modules.Geo.Domain/ValueObjects/Altitude.cs
@@ -1,3 +1,3 @@
 namespace GoActive.Modules.Geo.Domain.ValueObjects;
 
-public readonly record struct Altitude(double Value);
+public readonly record struct Altitude(float Value);
